added checks on Decoder.objCurrent to handle case if it's nil.
object (o) and group (g) lines are optional in OBJ. the parser doesn't create a group if
these lines are not present, which breaks parsing of face and usemtl lines. I added checks
to create 'default' groups when first encountering a usemtl or face line if no group or
object was created.